requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
gitpython
python-dotenv
requests
sentence-transformers
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-prometheus
opentelemetry-sdk==1.21.0
opentelemetry-exporter-prometheus==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
prometheus-client==0.16.0
langchain-huggingface
prometheus-fastapi-instrumentator
prometheus_client